Added backend linting and formatting, unified prettier and eslint
