+++ client/.eslintrc.json
@@ -3,13 +3,17 @@
         "browser": true,
         "es2021": true
     },
-    "extends": "airbnb",
+    "extends": [
+        "airbnb",
+        "plugin:prettier/recommended"
+    ],
     "parserOptions": {
         "ecmaVersion": "latest",
         "sourceType": "module"
     },
     "rules": {
         "react/react-in-jsx-scope": "off",
-        "react/jsx-uses-react": "off"
+        "react/jsx-uses-react": "off",
+        "prettier/prettier": "error"
     }
 }
