--- client/.prettierrc.json
@@ -1,5 +1,0 @@
-{
-    "singleQuote": true,
-    "jsxSingleQuote": true,
-    "tabWidth": 4
-}
