+++ server/package.json
@@ -4,19 +4,26 @@
   "description": "Server for the TCES CRM system.",
   "main": "index.js",
   "scripts": {
-    "test": "vitest"
+    "test": "vitest",
+    "lint": "npx eslint src/**/*.js",
+    "format": "prettier --write ."
   },
   "author": "Blueprint",
   "license": "MIT",
   "dependencies": {
     "dotenv": "^16.3.1",
+    "eslint": "^8.52.0",
     "express": "^4.18.2",
     "mysql2": "^3.6.2",
-    "sequelize": "^6.33.0",
     "pino": "^8.16.0",
-    "pino-http": "^8.5.0"
+    "pino-http": "^8.5.0",
+    "prettier": "^3.0.3",
+    "sequelize": "^6.33.0"
   },
   "devDependencies": {
+    "eslint-config-airbnb": "^19.0.4",
+    "eslint-config-prettier": "^9.0.0",
+    "eslint-plugin-prettier": "^5.0.1",
     "pino-pretty": "^10.2.3",
     "vitest": "^0.34.6"
   }
