+++ server/src/middlewares/logger/pino.js
@@ -1,17 +1,17 @@
-const pino = require('pino');
+const pino = require("pino");
 
 module.exports = pino({
-    transport: {
-        target: 'pino-pretty',
-        options: {
-            colorize: true,
-            ignore: 'pid,hostname',
-        },
+  transport: {
+    target: "pino-pretty",
+    options: {
+      colorize: true,
+      ignore: "pid,hostname",
     },
-    formatters: {
-        level(label) {
-            return { level: label.toUpperCase() };
-        },
+  },
+  formatters: {
+    level(label) {
+      return { level: label.toUpperCase() };
     },
-    timestamp: pino.stdTimeFunctions.isoTime,
+  },
+  timestamp: pino.stdTimeFunctions.isoTime,
 });
