+++ server/src/middlewares/logger/pino_http.js
@@ -1,16 +1,16 @@
-const pino_http = require('pino-http');
+const pinoHTTP = require("pino-http");
 
-module.exports = pino_http({
-    transport: {
-        target: 'pino-pretty',
-        options: {
-            colorize: true,
-            ignore: 'pid,hostname',
-        },
+module.exports = pinoHTTP({
+  transport: {
+    target: "pino-pretty",
+    options: {
+      colorize: true,
+      ignore: "pid,hostname",
     },
-    formatters: {
-        level(label) {
-            return { level: label.toUpperCase() };
-        },
+  },
+  formatters: {
+    level(label) {
+      return { level: label.toUpperCase() };
     },
+  },
 });
